Revert "mixed_paint::collection: add parts spinentry to columns"
Inability to get access to the embedded SpinButton (easy in Python) meant
that changes to the SpinButton's values could not be accessed in a way that
made the mixer interface useable.  So an alternative mechanism for allowing
mixed paints to be used as paint components in the mixer will be developed.

This reverts commit 73a58bd5262f04d9f5f98633a6a31c8c128aafc0.